HBASE-7899 Cell block building tools: Cell codec and means of iterating an objects Cells
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1449420 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-common/src/main/java/org/apache/hadoop/hbase/util/test/RedundantKVGenerator.java
@@ -205,7 +205,7 @@ public class RedundantKVGenerator {
       randomizer.nextBytes(family);
     }
 
-    long baseTimestamp = Math.abs(randomizer.nextLong()) / baseTimestampDivide;
+    long baseTimestamp = Math.abs(randomizer.nextInt()) / baseTimestampDivide;
 
     byte[] value = new byte[valueLength];
